streamlit/requirements.txt

streamlit
pandas
numpy
hdfs
fitz
streamlit_pdf_viewer
tika
pdfminer.six
pdf2image
pytesseract
Pillow